Updated login, signup and profile content
