+++ controllers/homeRoutes.js
@@ -1,57 +1,64 @@
 const router = require('express').Router();
-const { Appointments,Doctors,History,Patients,Schedules } = require('../models');
+const {
+  Appointments,
+  History,
+  Patients
+} = require('../models');
 const withAuth = require('../utils/auth');
 
 
-router.get('/', async (req, res) => {
-    
-    if (!req.session.loggedIn) {
-        res.redirect('/login');
-        return;
-      }
+router.get('/', withAuth, async (req, res) => {
 
-    try {
-      // Get all posts and JOIN with user data
-      console.log(req.session.user_id, "**********")
-      const postData = await Patients.findByPk(req.session.user_id, {
-        include: [
-          {
-            model: History,
+  if (!req.session.loggedIn) {
+    res.redirect('/login');
+    return;
+  }
 
-          },
-          {
-            model: Appointments,
+  try {
+    // Get all posts and JOIN with user data
+    console.log(req.session.user_id, "**********")
+    const patientData = await Patients.findByPk(req.session.user_id, {
+      attributes: { exclude: ['password'] },
+      include: [{
+          model: History,
+        },
+        {
+          model: Appointments,
+        }
+      ],
+    });
 
-          },
-          
-        ],
-      });
-  
-      //console.log("postData",postData);
-      // Serialize data so the template can read it
-      // const posts = postData.map((post) => post.get({ plain: true }));
-      const patient = postData.get({ plain: true });
+    console.log("patientData", patientData);
+    // Serialize data so the template can read it
+    // const posts = postData.map((post) => post.get({ plain: true }));
+    const patient = patientData.get({
+      plain: true
+    });
 
-      console.log("patient",patient);
-      // Pass serialized data and session flag into template
-      res.render('home', { 
-        patient, 
-        logged_in: req.session.logged_in 
-      });
-    } catch (err) {
-      console.log("err:",err);
-      res.status(500).json(err);
-    };
-  });
+    console.log("patient", patient);
+    // Pass serialized data and session flag into template
+    res.render('home', {
+      patient,
+      loggedIn: req.session.loggedIn
+    });
+  } catch (err) {
+    console.log("err:", err);
+    res.status(500).json(err);
+  };
+});
 
-  router.get('/login', (req, res) => {
-    // If the user is already logged in, redirect the request to another route
-    if (req.session.logged_in) {
-      res.redirect('/');
-      return;
-    };
-  
-    res.render('login');
-  });
+router.get('/login', (req, res) => {
+  // If the user is already logged in, redirect the request to another route
+  if (req.session.loggedIn) {
+    res.redirect('/');
+    return;
+  };
 
-  module.exports = router;
+  res.render('login');
+});
+
+router.get('/signup', (req, res) => {
+  res.render('signup');
+});
+
+module.exports = router;
