+++ public/css/style.css
@@ -1,0 +1,90 @@
+* {
+    box-sizing: border-box;
+}
+
+body {
+    margin: 0;
+    font-family: Arial, Helvetica, sans-serif;
+}
+
+.header {
+    overflow: hidden;
+    background-color: royalblue;
+    padding: 10px 5px;
+}
+
+.header a {
+    float: left;
+    color: white;
+    /* text-align: center; */
+    padding: 12px;
+    text-decoration: none;
+    font-size: 18px;
+    line-height: 25px;
+    border-radius: 4px;
+}
+
+.header a.logo {
+    font-size: 25px;
+    font-weight: bold;
+}
+
+.header a:hover {
+    /* background-color: #ddd; */
+    color: black;
+}
+
+/* .header a.active {
+    background-color: dodgerblue;
+    color: white;
+} */
+
+.header-right {
+    float: right;
+}
+
+.img-header {
+    width: 7%;
+    height: 7%;
+}
+
+img {
+    width: 30%;
+    height: 30%;
+}
+
+.card {
+    border: none;
+}
+
+.btn-color {
+    background-color: #0e1c36;
+    color: #fff;
+
+}
+
+.profile-image-pic {
+    height: 200px;
+    width: 200px;
+    object-fit: cover;
+}
+
+.cardbody-color {
+    background-color: #ebf2fa;
+}
+
+a {
+    text-decoration: none;
+}
+
+@media screen and (max-width: 500px) {
+    .header a {
+        float: none;
+        display: block;
+        text-align: left;
+    }
+
+    .header-right {
+        float: none;
+    }
+}
