+++ public/js/logout.js
@@ -1,5 +1,5 @@
 const logout = async () => {
-  const response = await fetch('/api/users/logout', {
+  const response = await fetch('/api/patientRoutes/logout', {
     method: 'POST',
     headers: { 'Content-Type': 'application/json' },
   });
