+++ views/layouts/main.handlebars
@@ -6,37 +6,45 @@
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <link rel='stylesheet' href='https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css' />
+  <link rel="stylesheet" href="/css/style.css">
   <title>Medical Portal!</title>
 </head>
 
-
 <body>
-<div class="flex-column justify-space-around">
+  <div class="flex-column justify-space-around">
     <header class="display-flex justify-space-between align-center p-2">
-      <h1>
-        <a href="/">⭐️ Medical Portal </a>
-      </h1>
-      <nav>
-        <!-- Conditionally render login or logout links -->
-        {{#if logged_in}}
-        <a href="/profile">profile</a> |
-        <button class="no-button" id="logout">logout</button>
-        {{else}}
-        <a href="/login">login</a>
-        {{/if}}
-      </nav>
+      <div class="header">
+
+        <a href="/" class="logo">
+          <img class="img-header" src="/image/healthcare.png" alt="Card appointment" />
+          Medical Portal
+        </a>
+        <div class="header-right">
+          {{!-- <a class="active" href="/">Medical Portal</a> --}}
+          <a href="#contact">Contact</a>
+          <a href="#about">About</a>
+
+          <!-- Conditionally render login or logout links -->
+          {{#if loggedIn}}
+          <a id="logout" href="#about">Logout</a>
+          {{else}}
+          <a href="/login">Login</a>
+          {{/if}}
+
+        </div>
+      </div>
     </header>
     <main class="container container-fluid mt-5">
       <!-- Render the sub layout -->
       {{{ body }}}
     </main>
     <footer>
-      
+
     </footer>
   </div>
 
   <!-- Render script for logged in users only -->
-  {{#if logged_in}}
+  {{#if loggedIn}}
   <script src="/js/logout.js"></script>
   {{/if}}
 </body>
