+++ views/login.handlebars
@@ -1,42 +1,29 @@
-<div class="row">
-  <div class="col-md-6">
-    <h2>Login</h2>
+<div class="container">
+  <div class="row">
+    <div class="col-md-6 offset-md-3">
+      {{!-- <h2 class="text-center text-dark mt-5">Login Form</h2>
+      <div class="text-center mb-5 text-dark">Made with bootstrap</div> --}}
+      <div class="card my-5">
+        <form id="login-form" class="card-body cardbody-color p-lg-5">
+          <div class="text-center">
+            <img src="/image/login_patient.png"
+              class="img-fluid profile-image-pic img-thumbnail rounded-circle my-3" width="150px" alt="profile">
+          </div>
+          <div class="mb-3">
+            <input type="text" class="form-control" id="email-login" aria-describedby="emailHelp" placeholder="Email">
+          </div>
+          <div class="mb-3">
+            <input type="password" class="form-control" id="password-login" placeholder="Password">
+          </div>
+          <div class="text-center"><button type="submit" class="btn btn-color px-5 mb-5 w-100">Login</button></div>
+          <div id="emailHelp" class="form-text text-center mb-5 text-dark">Not
+            Registered? <a href="/signup" class="login-link text-dark fw-bold">Sign up instead</a>
+          </div>
+        </form>
+      </div>
 
-    <form class="form login-form">
-      <div class="form-group">
-        <label for="email-login">email:</label>
-        <input class="form-input" type="text" id="email-login" />
-      </div>
-      <div class="form-group">
-        <label for="password-login">password:</label>
-        <input class="form-input" type="password" id="password-login" />
-      </div>
-      <div class="form-group">
-        <button class="btn btn-primary" type="submit">login</button>
-      </div>
-    </form>
-  </div>
-  <div class="col-md-6">
-    <h2>Signup</h2>
-
-    <form class="form signup-form">
-      <div class="form-group">
-        <label for="name-signup">name:</label>
-        <input class="form-input" type="text" id="name-signup" />
-      </div>
-      <div class="form-group">
-        <label for="email-signup">email:</label>
-        <input class="form-input" type="text" id="email-signup" />
-      </div>
-      <div class="form-group">
-        <label for="password-signup">password:</label>
-        <input class="form-input" type="password" id="password-signup" />
-      </div>
-      <div class="form-group">
-        <button class="btn btn-primary" type="submit">signup</button>
-      </div>
-    </form>
+    </div>
   </div>
 </div>
 
-<script src="./js/login.js"></script>
+<script src="/js/login.js"></script>
