--- views/profile.handlebars
@@ -1,26 +1,0 @@
-<div class="row">
-    <div class="col-md-6">
-        <h2> ⭐️  Medical Portal</h2>
-    </div>
-
-        <div class="card" style="width: 18rem;">
-    <div class="card-body">
-        <h5 class="card-title">Appointments</h5>
-        <h6 class="card-subtitle mb-2 text-muted">{{appointment.appointment_start}}</h6>
-        <p class="card-text">{{appointment.concerns}}</p>
-        <p class="card-text"><strong>{{appointment.status}}</strong></p>
-        <p class="card-text">{{doctors.name}}</p>
-    </div>
-    </div>
-
-
-        <div class="card" style="width: 18rem;">
-    <div class="card-body">
-        <h5 class="card-title">Patient History</h5>
-        <h6 class="card-subtitle mb-2 text-muted">Current Medications: {{history.medications}}</h6>
-        <p class="card-text">Previous History: {{history.data}}</p>
-    </div>
-    </div>
-
-
-</div>
