+++ views/signup.handlebars
@@ -1,0 +1,51 @@
+{{!-- <main class="form-signin text-center">
+  <form id="signup-form">
+    <h1 class="h3 mb-3 fw-normal">Sign up today!</h1>
+
+    <div class="form-floating mb-3">
+      <input type="text" class="form-control" id="username-signup" placeholder="Doe246" autofocus="true">
+      <label for="username-signup">Username</label>
+    </div>
+    <div class="form-floating mb-3">
+      <input type="password" class="form-control" id="password-signup" placeholder="Password" autofocus="true">
+      <label for="password-signup">Password</label>
+    </div>
+
+    <button class="w-100 btn btn-lg btn-info mb-3" type="submit">Sign up</button>
+    <a class="login-link" href="/login">Log in instead</a>
+  </form>
+</main>
+
+<script src="/js/signup.js"></script> --}}
+
+
+<div class="container">
+  <div class="row">
+    <div class="col-md-6 offset-md-3">
+      <div class="card my-5">
+        <form id="signup-form" class="card-body cardbody-color p-lg-5">
+          <div class="text-center">
+            <img src="/image/signup_patient.png"
+              class="img-fluid profile-image-pic img-thumbnail rounded-circle my-3" width="200px" alt="profile">
+          </div>
+          <div class="mb-3">
+            <input type="text" class="form-control" id="name-signup" aria-describedby="nameHelp" placeholder="Name">
+          </div>
+          <div class="mb-3">
+            <input type="text" class="form-control" id="email-signup" aria-describedby="emailHelp" placeholder="Email">
+          </div>
+          <div class="mb-3">
+            <input type="password" class="form-control" id="password-signup" placeholder="Password">
+          </div>
+          <div class="text-center"><button type="submit" class="btn btn-color px-5 mb-5 w-100">Sign up</button></div>
+          <div class="form-text text-center mb-5 text-dark">
+            <a href="/login" class="login-link text-dark fw-bold">Log in instead</a>
+          </div>
+        </form>
+      </div>
+
+    </div>
+  </div>
+</div>
+
+<script src="/js/signup.js"></script>
